Continue PR response, still wip
