+++ app/controllers/projects_controller.rb
@@ -10,7 +10,6 @@ class ProjectsController < ApplicationController
   before_action :require_can_edit_projects!, only: [:edit, :update]
   before_action :set_project, only: [:show, :update, :edit, :destroy]
   before_action :set_census_params, only: [:show]
-  before_action :set_location_params, only: [:show]
   before_action :require_can_view_confidential_project_names!, if: -> { !can_edit_projects? && @project.confidential? }
 
   include ArelHelper
@@ -61,45 +60,6 @@ class ProjectsController < ApplicationController
       aggregation_level: :by_project,
       aggregation_type: :inventory,
     }
-  end
-
-  private def set_location_params
-    return unless RailsDrivers.loaded.include?(:client_location_history)
-
-    @locations = @project.enrollment_location_histories.where(located_on: location_filter.range)
-    @markers = @locations.map { |l| l.as_marker(current_user, [:name, :seen_on]) }
-    @bounds = ClientLocationHistory::Location.bounds(@locations)
-    @options = {
-      bounds: @bounds,
-      cluster: true,
-      marker_color: ClientLocationHistory::Location::MARKER_COLOR,
-    }
-  end
-
-  private def location_filter
-    @location_filter ||= filter_class.new(
-      user_id: current_user.id,
-      enforce_one_year_range: false,
-    ).set_from_params(location_filter_params[:location_filters])
-  end
-
-  def location_filter_params
-    opts = params
-    opts[:location_filters] ||= {}
-    opts[:location_filters][:enforce_one_year_range] = false
-    opts[:location_filters][:start] ||= 1.year.ago
-    opts[:location_filters][:end] ||= Date.current
-    opts.permit(
-      location_filters: [
-        :start,
-        :end,
-      ],
-    )
-  end
-  helper_method :location_filter_params
-
-  private def filter_class
-    ::Filters::FilterBase
   end
 
   private def project_scope
