+++ app/views/projects/show.haml
@@ -121,10 +121,6 @@
   %p The following import overrides are in effect for this project.
   .mb-4.card
     = render 'hmis_csv_importer/import_overrides/table', overrides: overrides, data_source: @project.data_source, editable: false, show_associated_project: false
-
-- if RailsDrivers.loaded.include?(:client_location_history) && @project.can?(current_user, permission: :can_view_project_locations)
-  = render 'client_location_history/projects/map'
-
 - if @show_census
   = render 'census'
 = render 'funders', funders: @project.funders
