--- drivers/client_location_history/app/views/client_location_history/projects/_map.haml
@@ -1,13 +1,0 @@
-%h2= "Location Map for #{@project.name}"
-= render 'client_location_history/projects/filter'
-.mb-4.card
-  - if @locations.any?
-    #client__location-history
-    = content_for :page_js do
-      :javascript
-        var markers = #{@markers.to_json.html_safe};
-        var options = #{@options.to_json.html_safe};
-        var map = new App.Maps.MapWithMarkers('client__location-history', markers, options);
-  - else
-    .no-data
-      No locations
