+++ drivers/client_location_history/app/views/client_location_history/projects/map.haml
@@ -1,0 +1,14 @@
+- title = "Location Map for #{@project.name}"
+- content_for :title, title
+%h1= title
+= render 'filter'
+- if @locations.any?
+  #client__location-history
+  = content_for :page_js do
+    :javascript
+      var markers = #{@markers.to_json.html_safe};
+      var options = #{@options.to_json.html_safe};
+      var map = new App.Maps.MapWithMarkers('client__location-history', markers, options);
+- else
+  .no-data
+    No locations
