+++ drivers/client_location_history/config/routes.rb
@@ -9,6 +9,9 @@ BostonHmis::Application.routes.draw do
     resources :clients, only: [:none] do
       get :map, on: :member
     end
+    resources :projects, only: [:none] do
+      get :map, on: :member
+    end
     namespace :warehouse_reports do
       resources :client_location_history, only: [:index]
     end
